+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add CallableMetadata.returnType
 * Add ClassMetadata.isData
+* Add ConstructorMetadata.isPrimary
 * Update to Kotlin 1.2.0
 
 ## 1.0.0 (Feb 15, 2016)